Add Objects that point to "users"
A user can be the owner of the body. The use of this feature is to "tie" a
body to another object. Notable examples are some Entity object in an
implementing game. Having the pointer to the user allows an implementation
to figure out if additional logic should be performed on a body.

Signed-off-by: Chris Sanders <zidane@outlook.com>