Add ghost objects.
Ghost objects are "detectors". They will report collisions that occur but
won't incur an adjustment for the other body within the collision.
Callbacks are still alerted that collision occurred.

Examples of use are manual movement of players and movement in water
(where you don't want React to adjust the body out of the water).

Signed-off-by: Chris Sanders <zidane@outlook.com>